Implementation of Hypre GPU matrix assembly. Hypre assembly is working on up to 32 GPUs. Can do momentum as either scalar or vector (segregated) solver.
Also, added an option for Hypre assembly, ensure_reproducible, that when enabled, sorts
the matrix/rhs values that are summed for each entry, by absolute value floating point
ascending order. This ensures an ordering and a matrix/rhs pair that is guaranteed to be
the same given the same input (lhs/rhs). In spite of this, runs still aren't deterministic.
Sigh.